Repair missing workspace in call/inheritance hierarchy
With new multi workspace lsp-mode there is no single current workspace
associated with a buffer anymore. This leads to "Cquery is not enabled
in this buffer." errors in cquery-call-hierarchy and
cquery-inheritance-hierarchy.

Fix by adding a new utility function cquery--get-lsp-workspace to get
the current cquery workspace from the list of buffer workspaces and
apply during the affected calls.